requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.0
python-multipart==0.0.9